# docs/swagger.yaml
basePath: /
definitions:
  Agency.CreateAgency:
    properties:
      address:
        type: string
      city:
        type: string
      description:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      type:
        items:
          type: string
        type: array
      website:
        type: string
    required:
    - address
    - city
    - description
    - email
    - name
    - phone
    - type
    - website
    type: object
  Agency.UpdateAgency:
    properties:
      address:
        type: string
      city:
        type: string
      description:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      website:
        type: string
    type: object
  Company.CreateCompany:
    properties:
      address:
        type: string
      areas_of_activity:
        items:
          type: string
        type: array
      city:
        type: string
      description:
        type: string
      email:
        type: string
      image:
        $ref: '#/definitions/utils.Base64Struct'
      name:
        type: string
      phone:
        type: string
      website:
        type: string
    required:
    - email
    - image
    - name
    type: object
  Company.UpdateCompany:
    properties:
      address:
        type: string
      areas_of_activity:
        items:
          type: string
        type: array
      city:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: string
      image:
        $ref: '#/definitions/utils.Base64Struct'
      name:
        type: string
      phone:
        type: string
      website:
        type: string
    type: object
  channel.CreateChannel:
    properties:
      address:
        type: string
      description:
        type: string
      email:
        type: string
      fm_cpe_1:
        type: string
      fm_cpe_2:
        type: string
      fm_daily_avg_1:
        type: string
      fm_daily_avg_2:
        type: string
      fm_second_eval_1:
        type: string
      fm_second_eval_2:
        type: string
      image:
        $ref: '#/definitions/utils.Base64Struct'
      name:
        type: string
      phone:
        type: string
      tv_daily_avg_views:
        type: string
      tv_univision_number:
        type: string
      type:
        type: string
      website:
        type: string
    required:
    - image
    - name
    - type
    type: object
  channel.UpdateChannel:
    properties:
      address:
        type: string
      description:
        type: string
      email:
        type: string
      fm_cpe_1:
        type: string
      fm_cpe_2:
        type: string
      fm_daily_avg_1:
        type: string
      fm_daily_avg_2:
        type: string
      fm_second_eval_1:
        type: string
      fm_second_eval_2:
        type: string
      image:
        $ref: '#/definitions/utils.Base64Struct'
      name:
        type: string
      phone:
        type: string
      tv_daily_avg_views:
        type: string
      tv_univision_number:
        type: string
      type:
        type: string
      website:
        type: string
    required:
    - image
    - name
    - type
    type: object
  reference.CreateReference:
    properties:
      code:
        type: string
      description:
        type: string
      field1:
        type: string
      field2:
        type: string
      field3:
        type: string
      image:
        $ref: '#/definitions/utils.Base64Struct'
      name:
        type: string
    required:
    - code
    - name
    type: object
  reference.UpdateReference:
    properties:
      code:
        type: string
      description:
        type: string
      field1:
        type: string
      field2:
        type: string
      field3:
        type: string
      name:
        type: string
    required:
    - code
    - name
    type: object
...
  role.RoleUpdateInput:
    properties:
      description:
        type: string
      field1:
        type: string
      field2:
        type: string
      field3:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  user.SignInInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  user.SignUpCompany:
    properties:
      company_account:
        type: string
      email:
        type: string
      location:
        type: string
      manager_phone_number:
        type: string
      name:
        type: string
      password:
        type: string
      passwordConfirm:
        type: string
      phone_number:
        type: string
      photo:
        $ref: '#/definitions/utils.Base64Struct'
      prole_id:
        type: string
      role_id:
        type: string
    required:
    - email
    - name
    - password
    - passwordConfirm
    - phone_number
    type: object
  user.SignUpInfluencer:
    properties:
      audience_interests:
        type: string
      average_likes:
        type: number
      avg_comments:
        type: number
      avg_likes:
        type: number
      avg_reel_plays:
        type: number
      avg_views:
        type: number
      bio:
        type: string
      email:
        type: string
      engagement_rate:
        type: number
      followers:
        type: number
      gender_split:
        type: string
      ig_name:
        type: string
      location:
        type: string
      name:
        type: string
      password:
        type: string
      passwordConfirm:
        type: string
      phone_number:
        type: string
      photo:
        $ref: '#/definitions/utils.Base64Struct'
      popular_posts:
        type: string
      role_id:
        type: string
      total_posts:
        type: number
    required:
    - email
    - ig_name
    - name
    - password
    - passwordConfirm
    type: object
  user.SignUpInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      passwordConfirm:
        minLength: 8
        type: string
      photo:
        $ref: '#/definitions/utils.Base64Struct'
    required:
    - email
    - name
    - password
    - passwordConfirm
    type: object
...
  utils.RequestObj:
    properties:
      default_param:
        items:
          $ref: '#/definitions/utils.DefaultParam'
        type: array
      filter:
        items:
          $ref: '#/definitions/utils.FilterObj'
        type: array
      glob_operation:
        type: string
      page_no:
        type: integer
      per_page:
        type: integer
      sort:
        type: string
    type: object
  utils.ResponseObj:
    properties:
      data: {}
      error: {}
      pagination:
        $ref: '#/definitions/utils.Pagination'
      response_code:
        type: integer
      response_msg:
        type: string
    type: object
host: http://103.168.56.249:8080
info:
  contact: {}
  description: This is a sample API with Fiber and Swagger
  title: Ramen API
  version: "1.0"
paths:
  /agent/{id}:
    delete:
      consumes:
      - application/json
      description: Delete user
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - Agency
    get:
      consumes:
      - application/json
      description: Get agent
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Get agent
      tags:
      - Agency
    put:
      consumes:
      - application/json
      description: Update agent
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Agency
        in: body
        name: agency
        required: true
        schema:
          $ref: '#/definitions/Agency.UpdateAgency'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Update agent
      tags:
      - Agency
  /agent/create:
    post:
      consumes:
      - application/json
      description: Create a new agency
      parameters:
      - description: Agency
        in: body
        name: agency
        required: true
        schema:
          $ref: '#/definitions/Agency.CreateAgency'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Create a new agency
      tags:
      - Agency
  /agent/list:
    post:
      consumes:
      - application/json
      description: Get agent list
      parameters:
      - description: Agency
        in: body
        name: agency
        required: true
        schema:
          $ref: '#/definitions/utils.RequestObj'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Get agent list
      tags:
      - Agency
  /auth/login:
    post:
      consumes:
      - application/json
      description: Sign in user
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Sign in user
      tags:
      - Auth
  /auth/signup/admin:
    post:
      consumes:
      - application/json
      description: Create a new admin
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.SignUpInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Create a new admin
      tags:
      - Auth
  /auth/signup/company:
    post:
      consumes:
      - application/json
      description: Sign up company
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.SignUpCompany'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Sign up company
      tags:
      - Auth
  /auth/signup/influencer:
    post:
      consumes:
      - application/json
      description: Sign up influencer
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.SignUpInfluencer'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Sign up influencer
      tags:
      - Auth
  /channel:
    post:
      consumes:
      - application/json
      description: Create channel
      parameters:
      - description: CreateChannel
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/channel.CreateChannel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Create channel
      tags:
      - Channel
  /channel/{id}:
    delete:
      consumes:
      - application/json
      description: Delete channel
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Delete channel
      tags:
      - Channel
    get:
      consumes:
      - application/json
      description: Get channel
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Get channel
      tags:
      - Channel
    put:
      consumes:
      - application/json
      description: Update channel
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: UpdateChannel
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/channel.UpdateChannel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Update channel
      tags:
      - Channel
  /channel/list:
    post:
      consumes:
      - application/json
      description: Get channel list
      parameters:
      - description: RequestObj
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/utils.RequestObj'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Get channel list
      tags:
      - Channel
  /company:
    post:
      consumes:
      - application/json
      description: Create company
      parameters:
      - description: CreateCompany
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/Company.CreateCompany'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Create company
      tags:
      - Company
    put:
      consumes:
      - application/json
      description: Update company
      parameters:
      - description: UpdateCompany
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/Company.UpdateCompany'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Update company
      tags:
      - Company
  /company/{id}:
    delete:
      consumes:
      - application/json
      description: Delete company
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Delete company
      tags:
      - Company
    get:
      consumes:
      - application/json
      description: Get company
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Get company
      tags:
      - Company
  /company/list:
    post:
      consumes:
      - application/json
      description: Get company list
      parameters:
      - description: RequestObj
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/utils.RequestObj'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Get company list
      tags:
      - Company
  /file/{name}:
    get:
      consumes:
      - application/json
      description: Get file by name
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Get file by name
      tags:
      - File
  /reference:
    post:
      consumes:
      - application/json
      description: Create a new reference
      parameters:
      - description: Reference
        in: body
        name: reference
        required: true
        schema:
          $ref: '#/definitions/reference.CreateReference'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Create a new reference
      tags:
      - Reference
  /reference/{id}:
    delete:
      consumes:
      - application/json
      description: Delete reference
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Delete reference
      tags:
      - Reference
    get:
      consumes:
      - application/json
      description: Get reference
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Get reference
      tags:
      - Reference
    put:
      consumes:
      - application/json
      description: Update reference
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: UpdateReference
        in: body
        name: reference
        required: true
        schema:
          $ref: '#/definitions/reference.UpdateReference'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Update reference
      tags:
      - Reference
  /reference/list:
    post:
      consumes:
      - application/json
      description: Get reference list
      parameters:
      - description: RequestObj
        in: body
        name: reference
        required: true
        schema:
          $ref: '#/definitions/utils.RequestObj'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Get reference list
      tags:
      - Reference
  /resources:
    post:
      consumes:
      - application/json
      description: Create resources
      parameters:
      - description: CreateReference
        in: body
        name: resources
        required: true
        schema:
          $ref: '#/definitions/resources.CreateResources'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      summary: Create resources
      tags:
      - Resources
  /resources/{id}:
    delete:
      consumes:
      - application/json
      description: Delete resources
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      summary: Delete resources
      tags:
      - Resources
    get:
      consumes:
      - application/json
      description: Get resources
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      summary: Get resources
      tags:
      - Resources
    put:
      consumes:
      - application/json
      description: Update resources
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: UpdateReference
        in: body
        name: resources
        required: true
        schema:
          $ref: '#/definitions/resources.UpdateResources'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      summary: Update resources
      tags:
      - Resources
  /resources/list:
    post:
      consumes:
      - application/json
      description: List resources
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/utils.RequestObj'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      summary: List resources
      tags:
      - Resources
  /role/{id}:
    delete:
      consumes:
      - application/json
      description: Delete role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Delete role
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Update role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Role
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/role.RoleUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Update role
      tags:
      - Role
  /role/create:
    post:
      consumes:
      - application/json
      description: Create a new role
      parameters:
      - description: Role
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/role.RoleCreateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Create a new role
      tags:
      - Role
  /role/list:
    post:
      consumes:
      - application/json
      description: Get role list
      parameters:
      - description: Role
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/utils.RequestObj'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Get role list
      tags:
      - Role
  /users/{user_id}:
    delete:
      consumes:
      - application/json
      description: Delete user
      parameters:
      - description: ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: ID
        in: path
        name: user_id
        required: true
        type: string
      - description: UserUpdate
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - User
  /users/list:
    post:
      consumes:
      - application/json
      description: Get user list
      parameters:
      - description: RequestObj
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/utils.RequestObj'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseObj'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseObj'
      security:
      - ApiKeyAuth: []
      summary: Get user list
      tags:
      - User
  /users/me:
    get:
      consumes:
      - application/json
      description: Get user info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get user info
      tags:
      - User
swagger: "2.0"